modularizating code part2
